Remove redundant chat client dependency from ThreadController
